==> 1_data_processing.ipynb <==
## Trump Twitter

To analyze Trump's tweet I will be using a csv data containing tweets from 2009 to 2020. 

I will be performing the following tasks:
* Data Processing
    * Yearly Corpus
    * Document Term Matrix
* Exploratory Data Analysis
* Sentiment Analysis
    * TextBlob
* Topic Modelling
    * Latent Dirichlet Allocation
* Text Generation 
    * Markov's Chain 
    * Recurrent Neural Network

<b>Note:</b> Trump's tweet are only until June 2020.

## Data Processing

##### Importing Libraries

import pandas as pd
import matplotlib.pyplot as plt
import time
import re
import string
import nltk
import pickle

##### Loading Data

data = pd.read_csv('data/realdonaldtrump.csv')
data.head()

##### Converting date from string to timestamp

data = data.astype({'date':'datetime64[ns]'})
type(data['date'][0])

##### Excluding irrelevant data for this notebook 

# Excluding id, link, mentions, hashtags

data = data[['content', 'date', 'retweets', 'favorites']]
data.head()

### Data Cleaning

##### Round 1

To perform data science operations on data, it has to be cleaned first.

Cleaning of data involves the following:
    * Convert the text into lowercase
    * Removal of brackets
    * Removal of mentions
    * Removal of links
    * Removal of punctuation
    * Removal of numeric and alphanumeric words

def clean_text_round1(text):
    
    # Convert all tweets to lowercase
    text = text.lower()
    
    # removing small bracket
    text = re.sub('\(', '', text)
    text = re.sub('\)', '', text)
    
    # removing mentions
    text = re.sub('(?<=^|(?<=[^a-zA-Z0-9-_\.]))@(\s[A-Za-z]+[A-Za-z0-9_]+)', '', text)
    
    # removing links
    text = re.sub(r'http\S+', "", text)
    
    # removing links
    text = re.sub(r"'s", "", text)
    
    # replacing -- with one white-space 
    text = re.sub('--', ' ', text)
    
    #removing punctuation
    text = re.sub('[%s]' % re.escape(string.punctuation), '', text)
    
    #removing numeric and alpha-numeric characters
    text = re.sub('\w*\d\w*', '', text)
    
    return text

round1 = lambda x: clean_text_round1(x)

data_clean = data
data_clean['content'] = data['content'].apply(round1)
data_clean.head()

##### Round 2

Removing Non-English words from the tweets

words = set(nltk.corpus.words.words())

def clean_text_round2(text):
    text = " ".join(w for w in nltk.wordpunct_tokenize(text) if w.lower() in words or not w.isalpha())
    return text

round2 = lambda x: clean_text_round2(x)

data_clean['content'] = data_clean['content'].apply(round2)
data_clean.head()

##### Pickle Clean Data

data_clean.to_pickle('data/data_clean.pkl')

### Building Document Term Matrix

* Document Term Matrix is a matrix which describes the frequency of terms that occur in a collection of documents. 
* In DTM rows represents documents and columns reperesents terms.
* In this case, each tweets of a year is considered as a document. 
* To create a DTM, corpus is required which has been created in the above cell.
* This corpus is then used to create a Count Vectorizer which is then used to create DTM.
* Using CountVectorizer stop words can also be eliminated.

columns = ['year', 'transcript']
corpus_yearly = pd.DataFrame(columns=columns)

group = data_clean.groupby(data_clean.date.dt.year)

year_list = []
transcript_list = []

for year, group_data in group:
    text = ' '.join(group_data['content'])
    year_list.append(year)
    transcript_list.append(text)

corpus_yearly['year'] = year_list
corpus_yearly['transcript'] = transcript_list

corpus_yearly.head()

from sklearn.feature_extraction.text import CountVectorizer

cv = CountVectorizer(stop_words = 'english')
data_cv = cv.fit_transform(corpus_yearly.transcript)
dtm_yearly = pd.DataFrame(data_cv.toarray(), columns = cv.get_feature_names())
dtm_yearly.index = corpus_yearly.year
dtm_yearly

##### Pickle Document Term Matrix

dtm_yearly.to_pickle('data/dtm_yearly.pkl')

##### Pickle Count Vectorizer

pickle.dump(cv, open('data/cv.pkl', 'wb'))

### Building Yearly Corpus

A yearly corpus will be formed by joining all the tweets of every year. 

group = data_clean.groupby(data_clean.date.dt.year)

corpus_dict = {}

for year, group_data in group:
    text = ' '.join(group_data['content'])
    text = " ".join(w for w in nltk.wordpunct_tokenize(text) if w.lower() in words or not w.isalpha())
    corpus_dict[year] = [text]

corpus_yearly = pd.DataFrame.from_dict(corpus_dict).transpose()
corpus_yearly.columns = ['transcript']
corpus_yearly = corpus_yearly.sort_index()
corpus_yearly

##### Pickle yearly corpus

corpus_yearly.to_pickle('data/corpus_yearly.pkl')
