==> 4_topic_modeling.ipynb <==
## Topic Modeling

Topic Modelling is one of the popular technique in NLP  which is used to determine what topics are present in the given corpus. 

I will be using <b>Latent Dirichlet Allocation</b> abbreviated as <b>LDA</b> to perform topic modeling.

In topic modelling, the order does not matter, hence, <b>Document Term Matrix</b> is used instead of corpus. 

Process:  DTM --> Gensim(LDA) --> Topics

import pandas as pd
import matplotlib.pyplot as plt
import time
from datetime import datetime
import calendar
import nltk
import pickle

from gensim import matutils, models
import scipy.sparse

from nltk import word_tokenize, pos_tag
import nltk
nltk.download('punkt')
nltk.download('averaged_perceptron_tagger')

##### Loading Data

data_clean = pd.read_pickle('data/data_clean.pkl')
data_clean.head()

##### Loading Yearly DTM

dtm_yearly = pd.read_pickle('data/dtm_yearly.pkl')
dtm_yearly.head()

##### Building term document matrix

# Required a term document matrix which is transpose of document term matrix

tdm_yearly = dtm_yearly.transpose()
tdm_yearly.head()

### Attempt 1: With all the words

For Latent Drichelet Algorithm term document matrix has to be converted into sparse matrix then into a specific gensim corpus

# Creating sparse matrix using term document matrix
sparse_counts = scipy.sparse.csr_matrix(tdm_yearly)
sparse_counts.todense()

# Creating gensim corpus using sparse matrix
corpus = matutils.Sparse2Corpus(sparse_counts)

# Gensim does require dictionary of all the terms and their respective locations in term-document-matrix

cv = pickle.load(open("data/cv.pkl", "rb"))
id2word = dict((v,k) for k,v in cv.vocabulary_.items())

# Applying LDA with 2 topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=2, passes=10)
lda.print_topics()

# Applying LDA with 3 topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=3, passes=10)
lda.print_topics()

# Applying LDA with 4 topics
lda = models.LdaModel(corpus=corpus, id2word=id2word, num_topics=4, passes=10)
lda.print_topics()

### Attempt 2 - Nouns Only

corpus_yearly = pd.read_pickle('data/corpus_yearly.pkl')

# Function to tokenize a text and pull out nouns
def nouns(text):
    # tag NN is for nouns
    is_noun = lambda pos: pos[:2] == 'NN'
    tokenized = word_tokenize(text)
    all_nouns = [word for (word, pos) in pos_tag(tokenized) if is_noun(pos)]
    return ' '.join(all_nouns)

corpus_yearly_n = pd.DataFrame(corpus_yearly.transcript.apply(nouns))
corpus_yearly_n

from sklearn.feature_extraction import text
from sklearn.feature_extraction.text import CountVectorizer

add_stop_words = ['best', 'good', 'true', 'man', 'today', 'bad', 'today','trump','great','thanks','thank', 'just','president','people','make','people','new','time','like','think','country','big']

stop_words = text.ENGLISH_STOP_WORDS.union(add_stop_words)

# re-create a document term matrix with only nouns

cv_n = CountVectorizer(stop_words = stop_words)
data_cv_n = cv_n.fit_transform(corpus_yearly_n.transcript)
dtm_yearly_n = pd.DataFrame(data_cv_n.toarray(), columns = cv_n.get_feature_names())
dtm_yearly_n.index = corpus_yearly_n.index
dtm_yearly_n

# create gensim corpus
corpus_n = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(dtm_yearly_n.transpose()))

# creating vocab dict
id2word_n = dict((v,k) for k, v in cv_n.vocabulary_.items())

# models

lda_n = models.LdaModel(corpus=corpus_n, id2word=id2word_n ,num_topics=2, passes=10)
lda_n.print_topics()

# models

lda_n = models.LdaModel(corpus=corpus_n, id2word=id2word_n ,num_topics=3, passes=10)
lda_n.print_topics()

# models

lda_n = models.LdaModel(corpus=corpus_n, id2word=id2word_n ,num_topics=4, passes=10)
lda_n.print_topics()

### Attempt 3 - Nouns and Adjectives

# Function to tokenize a text and pull out nouns and adjectives
def nouns_adjective(text):
    # tag NN is for nouns
    is_noun_adj = lambda pos: pos[:2] == 'NN' or pos[:2] == 'JJ'
    tokenized = word_tokenize(text)
    nouns_adj = [word for (word, pos) in pos_tag(tokenized) if is_noun_adj(pos)]
    return ' '.join(nouns_adj)

corpus_yearly_na = pd.DataFrame(corpus_yearly.transcript.apply(nouns_adjective))
corpus_yearly_na

cv_na = CountVectorizer(stop_words = stop_words)
data_cv_na = cv_na.fit_transform(corpus_yearly_na.transcript)
dtm_yearly_na = pd.DataFrame(data_cv_na.toarray(), columns = cv_na.get_feature_names())
dtm_yearly_na.index = corpus_yearly_na.index
dtm_yearly_na

# create gensim corpus
corpus_na = matutils.Sparse2Corpus(scipy.sparse.csr_matrix(dtm_yearly_na.transpose()))

# creating vocab dict
id2word_na = dict((v,k) for k, v in cv_na.vocabulary_.items())

# models

lda_na = models.LdaModel(corpus=corpus_na, id2word=id2word_na ,num_topics=2, passes=10)
lda_na.print_topics()

# models

lda_na = models.LdaModel(corpus=corpus_na, id2word=id2word_na ,num_topics=3, passes=10)
lda_na.print_topics()

# models

lda_na = models.LdaModel(corpus=corpus_na, id2word=id2word_na ,num_topics=4, passes=10)
lda_na.print_topics()

### Topics per year

lda_na = models.LdaModel(corpus=corpus_na, id2word=id2word_na ,num_topics=10, passes=80)
lda_na.print_topics()

lda_na.print_topics()[0]

corpus_transformed = lda_na[corpus_na]
year=2008
for a in corpus_transformed:
    year=year+1
    print(year)
    for i in a:
        print(lda_na.print_topics()[i[0]])
    print('')

### Results

<b></b>
<b>Attempt 1 - With all the words</b>
* Using all the words does not give results which are good enough. 

<b>Attempt 2 - Only Nouns</b>
* Using just nouns does not give results which are good enough. 

<b>Attempt 3 - Only Nouns and Adjectives </b>
* <b>2009 Topics: </b>
    * <b>7:</b> celebrity, money
    * <b>9:</b> china, business
    
* <b>2010 Topics: </b>
    * <b>7:</b> luck, celebrity, money
    
* <b>2011 Topics: </b>
    * <b>4:</b> Fake news, China, border, military
    * <b>9:</b> china, business
    
* <b>2012 Topics: </b>
    * <b>4:</b> Fake news, China, border, military
    * <b>9:</b> china, business
    
* <b>2013 Topics: </b>
    * <b>7:</b> luck, celebrity, money
    * <b>9:</b> china, business
    
* <b>2014 Topics: </b>
    * <b>9:</b> china, business
    
* <b>2015 Topics: </b>
    * <b>2:</b> vote, media, campaign
    * <b>9:</b> china, business
    
* <b>2016 Topics: </b>
    * <b>2:</b> vote, media, campaign
    * <b>4:</b> Fake news, China, border, military
    
* <b>2017 Topics: </b>
    * <b>0:</b> Fake News, tax, media, election, military
    * <b>4:</b> Fake news, China, border, military
    
* <b>2018 Topics: </b>
    * <b>0:</b> Fake News, tax, media, election, military
    * <b>2:</b> vote, media, campaign
    * <b>4:</b> Fake news, China, border, military
    
* <b>2019 Topics: </b> 
    * <b>4:</b> Fake news, China, border, military
    
* <b>2020 Topics: </b>
    * <b>4:</b> Fake news, China, border, military
