==> 6_text_generation_RNN.ipynb <==
### Text Generation using RNN

##### Importing Libraries

import tensorflow as tf
import keras
from tensorflow.keras.preprocessing.sequence import pad_sequences
from tensorflow.keras.layers import Embedding, LSTM, Dense, Bidirectional
from tensorflow.keras.preprocessing.text import Tokenizer
from tensorflow.keras.models import Sequential
from tensorflow.keras.optimizers import Adam
import numpy as np 

##### Loading Data

data_clean_path = 'data_clean.pkl'

import pandas as pd
data_clean = pd.read_pickle(data_clean_path)
data_clean.head()

##### Seperating 2020 tweets

data = ''

for tweet in data_clean.content[40939:]:
    data += '\n' + tweet

len(data)

##### Tokenizing data

from tensorflow.keras.preprocessing.text import Tokenizer

tokenizer = Tokenizer()

corpus = data.lower().split("\n")

tokenizer.fit_on_texts(corpus)
total_words = len(tokenizer.word_index) + 1

print(total_words) # no of unique words
print(tokenizer.word_index)

##### Create predictors and labels 

input_sequences = []
for line in corpus:
    token_list = tokenizer.texts_to_sequences([line])[0]
    for i in range(1, len(token_list)):
        n_gram_sequence = token_list[:i+1]
        input_sequences.append(n_gram_sequence)

# pad sequences 
max_sequence_len = max([len(x) for x in input_sequences])
input_sequences = np.array(pad_sequences(input_sequences, maxlen=max_sequence_len, padding='pre'))

# create predictors and label
xs, labels = input_sequences[:,:-1],input_sequences[:,-1]

ys = tf.keras.utils.to_categorical(labels, num_classes=total_words)

##### Building and Training the RNN model

model = Sequential()
model.add(Embedding(total_words, 64, input_length=max_sequence_len-1))
model.add(Bidirectional(LSTM(20)))
model.add(Dense(total_words, activation='softmax'))
model.compile(loss='categorical_crossentropy', optimizer='adam', metrics=['accuracy'])
history = model.fit(xs, ys, epochs=500, verbose=1)

##### Save the model locally

from keras.models import load_model

model.save('model/tweet_generation.h5')

##### Load the saved model

model_trump_tweet_generaation = tf.keras.models.load_model('model/tweet_generation.h5')

### Results

* From the model trained above, it is clear that the accuracy is 51.47%
* Accuracy is quite low and the tweet generated by this model might not be structurally or grammatically correct. 

##### Predicting tweet with some input

def predict_tweet(seed_text, next_words):
    for _ in range(next_words):
        token_list = tokenizer.texts_to_sequences([seed_text])[0]
        token_list = pad_sequences([token_list], maxlen=max_sequence_len-1, padding='pre')
        predicted = model_trump_tweet_generaation.predict_classes(token_list, verbose=0)
        output_word = ""
        for word, index in tokenizer.word_index.items():
            if index == predicted:
                output_word = word
                break
        seed_text += " " + output_word
    return seed_text

print("Original Tweet:\n" + data_clean.content[28054])
print("\nGenerated Tweet:\n" + predict_tweet(" ".join(data_clean.content[28054].split()[:5]), 9))

print("Original Tweet:\n" + data_clean.content[12608])
print("\nGenerated Tweet:\n" + predict_tweet(" ".join(data_clean.content[12608].split()[:5]), 9))

print("Original Tweet:\n" + data_clean.content[35543])
print("\nGenerated Tweet:\n" + predict_tweet(" ".join(data_clean.content[35543].split()[:5]), 13))

# 20270
print("Original Tweet:\n" + data_clean.content[20270])
print("\nGenerated Tweet:\n" + predict_tweet(" ".join(data_clean.content[20270].split()[:5]), 14))

# 12048
print("Original Tweet:\n" + data_clean.content[12048])
print("\nGenerated Tweet:\n" + predict_tweet(" ".join(data_clean.content[12048].split()[:5]), 14))
